+++ azext_iot/central/models/v1_1_preview/template.py
@@ -71,3 +71,9 @@ class Template(BaseTemplate):
                 self.id
             )
             raise CLIError(details)
+
+    def get_id_key(self):
+        return "@id"
+
+    def get_type_key(self):
+        return "@type"